refactor: global 설정의 database layer 추가 및 경로 리팩토링
